requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
flask>=3.0.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
